+++ dotnet/src/Connectors/Connectors.AzureAIInference/Extensions/AzureAIInferenceServiceCollectionExtensions.cs
@@ -152,75 +152,6 @@ public static class AzureAIInferenceServiceCollectionExtensions
         });
     }
 
-    /// <summary>
-    /// Add an Azure AI Inference <see cref="IEmbeddingGenerator"/> to the <see cref="IServiceCollection"/>.
-    /// </summary>
-    public static IServiceCollection AddAzureAIInferenceEmbeddingGenerator(
-        this IServiceCollection services,
-        string modelId,
-        string? apiKey = null,
-        Uri? endpoint = null,
-        HttpClient? httpClient = null,
-        string? serviceId = null,
-        string? openTelemetrySourceName = null,
-        Action<OpenTelemetryEmbeddingGenerator<string, Embedding<float>>>? openTelemetryConfig = null)
-    {
-        Verify.NotNull(services);
-        return services.AddKeyedSingleton<IEmbeddingGenerator<string, Embedding<float>>>(serviceId, (serviceProvider, _) =>
-        {
-            httpClient ??= serviceProvider.GetService<HttpClient>();
-            var options = ChatClientCore.GetClientOptions(httpClient);
-            var loggerFactory = serviceProvider.GetService<ILoggerFactory>();
-
-            var builder = new Azure.AI.Inference.EmbeddingsClient(endpoint, new Azure.AzureKeyCredential(apiKey ?? SingleSpace), options)
-                .AsIEmbeddingGenerator(modelId).AsBuilder();
-
-            if (loggerFactory is not null)
-            {
-                builder.UseLogging(loggerFactory).Build();
-            }
-
-            builder.UseOpenTelemetry(loggerFactory, openTelemetrySourceName, openTelemetryConfig);
-
-            return builder.Build();
-        });
-    }
-
-    /// <summary>
-    /// Add an Azure AI Inference <see cref="IEmbeddingGenerator"/> to the <see cref="IServiceCollection"/>.
-    /// </summary>
-    public static IServiceCollection AddAzureAIInferenceEmbeddingGenerator(
-        this IServiceCollection services,
-        string modelId,
-        TokenCredential credential,
-        Uri? endpoint = null,
-        HttpClient? httpClient = null,
-        string? serviceId = null,
-        string? openTelemetrySourceName = null,
-        Action<OpenTelemetryEmbeddingGenerator<string, Embedding<float>>>? openTelemetryConfig = null)
-    {
-        Verify.NotNull(services);
-        return services.AddKeyedSingleton<IEmbeddingGenerator<string, Embedding<float>>>(serviceId, (serviceProvider, _) =>
-        {
-            httpClient ??= serviceProvider.GetService<HttpClient>();
-            var options = ChatClientCore.GetClientOptions(httpClient);
-
-            var loggerFactory = serviceProvider.GetService<ILoggerFactory>();
-            var builder = new Azure.AI.Inference.EmbeddingsClient(endpoint, credential, options)
-                .AsIEmbeddingGenerator(modelId)
-                .AsBuilder();
-
-            if (loggerFactory is not null)
-            {
-                builder.UseLogging(loggerFactory).Build();
-            }
-
-            builder.UseOpenTelemetry(loggerFactory, openTelemetrySourceName, openTelemetryConfig);
-
-            return builder.Build();
-        });
-    }
-
     #region Private
 
     /// <summary>
